+++ scriptconfig/config.py
@@ -87,12 +87,12 @@ import copy
 import io
 import json
 import numpy as np
-from .dict_like import DictLike
-from . import smartcast
-from .file_like import FileLike
-from .value import Value
+from scriptconfig.dict_like import DictLike
+from scriptconfig import smartcast
+from scriptconfig.file_like import FileLike
+from scriptconfig.value import Value
 
-__all__ = ['Config']
+__all__ = ['Config', 'define']
 
 
 def scfg_isinstance(item, cls):
@@ -107,13 +107,76 @@ def scfg_isinstance(item, cls):
         return isinstance(item, cls)
 
 
+def define(default={}, name=None):
+    """
+    Alternate method for defining a custom Config type
+    """
+    import uuid
+    if name is None:
+        hashid = str(uuid.uuid4()).replace('-', '_')
+        name = 'Config_{}'.format(hashid)
+    from textwrap import dedent
+    vals = {}
+    code = dedent(
+        '''
+        import scriptconfig as scfg
+        class {name}(scfg.Config):
+            pass
+        '''.strip('\n').format(name=name))
+    exec(code, vals)
+    cls = vals[name]
+    return cls
+
+
 class Config(ub.NiceRepr, DictLike):
     """
+    Base class for custom configuration objects
+
     A configuration that can be specified by commandline args, a yaml config
     file, and / or a in-code dictionary. To use, define a class variable named
     "default" and assing it to a dict of default values. You can also use
     special `Value` classes to denote types. You can also define a method
     `normalize`, to postprocess the arguments after this class receives them.
+
+    Usage:
+        Create a class that herits from this class.
+
+        Assign the "default" class-level variable as a dictionary of options
+
+        The keys of this dictionary must be command line friendly strings.
+
+        The values of the "defaults dictionary" can be literal values or
+        instances of the :class:`scriptconfig.Value` class, which allows
+        for specification of default values, type information, help strings,
+        and aliases.
+
+        You may also implement normalize (function with that takes no args and
+        has no return) to postprocess your results after initialization.
+
+        When creating an instance of the class the defaults variable is used
+        to make a dictionary-like object. You can override defaults by
+        specifying the ``data`` keyword argument to either a file path or
+        another dictionary. You can also specify ``cmdline=True`` to allow
+        the contents of ``sys.argv`` to influence the values of the new
+        object.
+
+        An instance of the config class behaves like a dictinary, except that
+        you cannot set keys that do not already exist (as specified in the
+        defaults dict).
+
+        Key Methods:
+
+            * dump - dump a json representation to a file
+
+            * dumps - dump a json representation to a string
+
+            * argparse - create the argparse object associated with this config
+
+            * argparse - create an :class:`argparse.ArgumentParser` object that
+                is defined by the defaults of this config.
+
+            * load - rewrite the values based on a filepath, dictionary, or
+                command line contents.
 
     Attributes:
         _data : this protected variable holds the raw state of the config
@@ -124,9 +187,10 @@ class Config(ub.NiceRepr, DictLike):
 
     Example:
         >>> # Inherit from `Config` and assign `default`
-        >>> class MyConfig(Config):
+        >>> import scriptconfig as scfg
+        >>> class MyConfig(scfg.Config):
         >>>     default = {
-        >>>         'option1': Value((1, 2, 3), tuple),
+        >>>         'option1': scfg.Value((1, 2, 3), tuple),
         >>>         'option2': 'bar',
         >>>         'option3': None,
         >>>     }
@@ -350,6 +414,9 @@ class Config(ub.NiceRepr, DictLike):
             # be Value objects, but this gets messy when the "default"
             # constructor argument is used. We should refactor so _data and
             # _default only store the raw current values, post-casting.
+            if key not in self.default:
+                # probably an alias
+                continue
             template = self.default[key]
             if not isinstance(template, Value):
                 # smartcast non-valued params from commandline
@@ -474,13 +541,15 @@ class Config(ub.NiceRepr, DictLike):
                 parser is updated with options from this config.
 
             special_options (bool, default=False):
-                adds special scriptconfig options.
+                adds special scriptconfig options, namely: --config, --dumps,
+                and --dump.
 
         Returns:
             argparse.ArgumentParser : a new or updated argument parser
 
         CommandLine:
-            xdoctest -m scriptconfig.config Config.argparse
+            xdoctest -m scriptconfig.config Config.argparse:0
+            xdoctest -m scriptconfig.config Config.argparse:1
 
         Example:
             >>> # You can now make instances of this class
@@ -495,13 +564,15 @@ class Config(ub.NiceRepr, DictLike):
             >>> class MyConfig(scfg.Config):
             >>>     description = 'my CLI description'
             >>>     default = {
-            >>>         'path1':  scfg.Value(None, position=1),
-            >>>         'path2':  scfg.Value(None, position=2),
+            >>>         'path1':  scfg.Value(None, position=1, alias='src'),
+            >>>         'path2':  scfg.Value(None, position=2, alias='dst'),
             >>>         'dry':  scfg.Value(False, isflag=True),
-            >>>         'approx':  scfg.Value(False, isflag=False),
+            >>>         'approx':  scfg.Value(False, isflag=False, alias=['a1', 'a2']),
             >>>     }
             >>> self = MyConfig()
-            >>> parser = self.argparse(special_options=True)
+            >>> special_options = True
+            >>> parser = None
+            >>> parser = self.argparse(special_options=special_options)
             >>> parser.print_help()
             >>> self._read_argv(argv=['objection', '42', '--path1=overruled!'])
             >>> print('self = {!r}'.format(self))
@@ -528,11 +599,12 @@ class Config(ub.NiceRepr, DictLike):
                 setattr(namespace, action.dest, values)
                 parser._explicitly_given.add(action.dest)
 
+        # IRC: this ensures each key has a real Value class
         _metadata = {
             key: self._data[key]
             for key, value in self._default.items()
             if isinstance(self._data[key], Value)
-        }
+        }  # :type: Dict[str, Value]
         _positions = {k: v.position for k, v in _metadata.items()
                       if v.position is not None}
         if _positions:
@@ -543,7 +615,42 @@ class Config(ub.NiceRepr, DictLike):
         else:
             _keyorder = list(self._default.keys())
 
+        def _add_arg(parser, name, key, argkw, positional, isflag, isalias):
+            _argkw = argkw.copy()
+
+            if isalias:
+                _argkw['help'] = 'alias of {}'.format(key)
+                _argkw.pop('default', None)
+            elif positional:
+                parser.add_argument(name, **_argkw)
+
+            if isflag:
+                # Can we support both flag and setitem methods of cli
+                # parsing?
+                if not isinstance(_argkw['default'], bool):
+                    raise ValueError('can only use isflag with bools')
+                _argkw.pop('type', None)
+                _argkw.pop('choices', None)
+                _argkw.pop('action', None)
+                _argkw.pop('nargs', None)
+                _argkw['dest'] = key
+
+                _argkw_true = _argkw.copy()
+                _argkw_true['action'] = 'store_true'
+
+                _argkw_false = _argkw.copy()
+                _argkw_false['action'] = 'store_false'
+                _argkw_false.pop('help', None)
+
+                parser.add_argument('--' + name, **_argkw_true)
+                parser.add_argument('--no-' + name, **_argkw_false)
+            else:
+                parser.add_argument('--' + name, **_argkw)
+
+        alias_registry = []
         for key, value in self._default.items():
+            # key: str
+            # value: Any | Value
             argkw = {}
             argkw['help'] = ''
             positional = None
@@ -555,30 +662,37 @@ class Config(ub.NiceRepr, DictLike):
                 value = _value.value
                 isflag = _value.isflag
                 positional = _value.position
+            else:
+                _value = value if isinstance(value, Value) else None
+
             if not argkw['help']:
                 argkw['help'] = '<undocumented>'
+
             argkw['default'] = value
             argkw['action'] = ParseAction
 
-            if positional:
-                parser.add_argument(key, **argkw)
-                parser.add_argument('--' + key, **argkw)
-            else:
-                if isflag:
-                    if not isinstance(argkw['default'], bool):
-                        raise ValueError('can only use isflag with bools')
-                    argkw.pop('type', None)
-                    argkw.pop('choices', None)
-                    argkw.pop('action', None)
-                    argkw.pop('nargs', None)
-                    argkw['dest'] = key
-                    parser.add_argument(
-                        '--' + key, action='store_true', **argkw)
-                    argkw.pop('help')
-                    parser.add_argument(
-                        '--no-' + key, action='store_false', **argkw)
-                else:
-                    parser.add_argument('--' + key, **argkw)
+            name = key
+            _add_arg(parser, name, key, argkw, positional, isflag, isalias=False)
+
+            mode = 1
+
+            if _value is not None:
+                if _value.alias:
+                    alts = _value.alias
+                    alts = alts if ub.iterable(alts) else [alts]
+                    for alias in alts:
+                        tup = (alias, key, argkw)
+                        alias_registry.append(tup)
+                        if mode == 0:
+                            name = alias
+                            _add_arg(parser, name, key, argkw, positional, isflag, isalias=True)
+
+        if mode == 1:
+            for tup in alias_registry:
+                (alias, key, argkw) = tup
+                name = alias
+                dest = key
+                _add_arg(parser, name, dest, argkw, positional, isflag, isalias=True)
 
         if special_options:
             parser.add_argument('--config', default=None, help=ub.codeblock(